streamlit/app/requirements.txt
streamlit==1.32.0
snowflake-connector-python==3.7.0
pandas==2.2.0
plotly==5.18.0
numpy==1.26.4
python-dotenv==1.0.1
requests==2.31.0
pytest==8.0.0
black==24.2.0
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
types-requests==2.31.0.20240125
types-python-dateutil==2.8.19.14 